refactor(fused_moe): split runtime dataclasses
Move fused MoE runtime argument structures into a dedicated dataclass package and update the routed experts, quantization paths, LoRA helpers, and tests to use the new entrypoints.

This removes the legacy stage contract modules and keeps the fusion-config lookup centralized in moe_utils.
